ZIOfy `uuid4cats-effect` lib
