+++ .jvmopts
@@ -1,0 +1,7 @@
+# This is a comment
+-Dfile.encoding=UTF8
+-Xms256M
+-Xmx8G
+-XX:+UseG1GC
+--add-opens=java.base/java.util=ALL-UNNAMED
+--add-opens=java.base/java.lang=ALL-UNNAMED
