+++ .jvmopts-ci
@@ -1,0 +1,8 @@
+# See GHA machines spec: https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources
+-Dfile.encoding=UTF8
+-Xms5G
+-Xmx5G
+-XX:+UseG1GC
+--add-opens=java.base/java.util=ALL-UNNAMED
+--add-opens=java.base/java.lang=ALL-UNNAMED
+-Dquill.macro.log=false # See https://getquill.io/#logging-logging-to-a-file-at-compile-time
