--- .mergify.yml
@@ -1,11 +1,0 @@
-pull_request_rules:
-  - name: Automatic merge of scala-steward PR
-    conditions:
-      - author=scala-steward
-      - check-success=Build and Test (ubuntu-latest, 2.13.11, temurin@8, rootJVM) 
-      - check-success=Build and Test (ubuntu-latest, 2.13.11, temurin@8, rootJVM) 
-      - check-success=Build and Test (ubuntu-latest, 3.3.0, temurin@8, rootJVM) 
-      - check-success=Build and Test (ubuntu-latest, 3.3.0, temurin@8, rootJS) 
-    actions:
-      merge:
-        method: merge
