+++ .scalafix.conf
@@ -1,0 +1,30 @@
+rules = [
+  Disable
+  DisableSyntax
+  ExplicitResultTypes
+  LeakingImplicitClassVal
+  NoAutoTupling
+  NoValInForComprehension
+  ProcedureSyntax
+  RemoveUnused
+  MissingFinal
+  EmptyCollectionsUnified
+]
+
+Disable {
+  ifSynthetic = [
+    "scala/Option.option2Iterable"
+    "scala/Predef.any2stringadd"
+  ]
+}
+
+DisableSyntax.regex = []
+
+RemoveUnused {
+  imports = true
+}
+
+DisableSyntax.noReturns = true
+DisableSyntax.noXml = true
+DisableSyntax.noFinalize = true
+DisableSyntax.noValPatterns = true
