+++ .scalafmt.conf
@@ -1,8 +1,30 @@
-version = 3.7.12
-runner.dialect = scala213
+version = "3.7.12"
+runner.dialect = Scala213Source3 # https://scalameta.org/scalafmt/docs/configuration.html#scala-2-with--xsource3
+maxColumn = 140
+align.preset = most
+continuationIndent.defnSite = 2
+assumeStandardLibraryStripMargin = true
+docstrings.style = Asterisk
+lineEndings = preserve
+includeCurlyBraceInSelectChains = true
+danglingParentheses.preset = true
+optIn.annotationNewlines = true
+newlines.alwaysBeforeMultilineDef = false
+newlines.implicitParamListModifierPrefer = before
+trailingCommas = multiple
+docstrings.wrap = no
 
-fileOverride {
-  "glob:**/scala-3/**" {
-    runner.dialect = scala3
-  }
+rewrite.rules = [RedundantBraces, SortModifiers]
+
+rewrite.sortModifiers.order = [
+  "implicit", "override", "private", "protected", "final", "sealed", "abstract", "lazy"
+]
+rewrite.redundantBraces.generalExpressions = false
+rewrite.redundantBraces.stringInterpolation = true
+rewriteTokens = {
+  "⇒": "=>"
+  "→": "->"
+  "←": "<-"
 }
+
+project.excludePaths = []
