--- CODE_OF_CONDUCT.md
@@ -1,14 +1,0 @@
-# Code of Conduct
-
-We are committed to providing a friendly, safe and welcoming environment for all, regardless of level of experience, gender, gender identity and expression, sexual orientation, disability, personal appearance, body size, race, ethnicity, age, religion, nationality, or other such characteristics.
-
-Everyone is expected to follow the [Scala Code of Conduct] when discussing the project on the available communication channels.
-
-
-## Moderation
-
-If you have any questions, concerns, or moderation requests, please contact a member of the project.
-
-- [Antoine Comte](mailto:antoine@comte.cc)
-
-[Scala Code of Conduct]: https://scala-lang.org/conduct/
