--- NOTICE
@@ -1,3 +1,0 @@
-uuid4cats
-Copyright 2023 Ant8e
-Licensed under Apache License 2.0 (see LICENSE)
