+++ Readme.md
@@ -1,35 +1,38 @@
+[//]: # (This file was autogenerated using `zio-sbt-website` plugin via `sbt generateReadme` command.)
+[//]: # (So please do not edit it manually. Instead, change "docs/index.md" file or sbt setting keys)
+[//]: # (e.g. "readmeDocumentation" and "readmeSupport".)
 
-[![Maven Central](https://maven-badges.herokuapp.com/maven-central/tech.ant8e/uuid4cats-effect_2.13/badge.svg)](https://maven-badges.herokuapp.com/maven-central/tech.ant8e/uuid4cats-effect_2.13)
-[![Code of Conduct](https://img.shields.io/badge/Code%20of%20Conduct-Scala-blue.svg)](CODE_OF_CONDUCT.md)
-![](https://github.com/ant8e/uuid4cats-effect/actions/workflows/ci.yml/badge.svg)
-[![codecov](https://codecov.io/gh/ant8e/uuid4cats-effect/branch/main/graph/badge.svg?token=QEUSQ3T053)](https://codecov.io/gh/ant8e/uuid4cats-effect)
-# uuid4cats-effect -  UUID and TypeID Generation for cats effect
+# zio-uuid
 
+[ZIO UUID](https://github.com/guizmaii/zio-uuid) is a "ziofied" fork
+of [uuid4cats-effect](https://github.com/ant8e/uuid4cats-effect) by [Antoine Comte](https://github.com/ant8e)
 
-Although cats-effect has some support for [generating UUIDs](https://typelevel.org/cats-effect/api/3.x/cats/effect/std/UUIDGen.html), it is limited to the UUIDv4 pseudo-random type.
-                                                                                 
+[![Production Ready](https://img.shields.io/badge/Project%20Stage-Production%20Ready-brightgreen.svg)](https://github.com/zio/zio/wiki/Project-Stages) ![CI Badge](https://github.com/zio/zio-uuid/workflows/CI/badge.svg) [![Sonatype Snapshots](https://img.shields.io/nexus/s/https/oss.sonatype.org/dev.zio/zio-uuid_2.13.svg?label=Sonatype%20Snapshot)](https://oss.sonatype.org/content/repositories/snapshots/dev/zio/zio-uuid_2.13/) [![zio-uuid](https://img.shields.io/github/stars/zio/zio-uuid?style=social)](https://github.com/zio/zio-uuid)
+
+## Introduction
 
 This library add support for the following types:
 
 |         |         time-based         | sortable | random |
 |--------:|:--------------------------:|:--------:|:------:|
 | UUID v1 | ✅ <br/> gregorian calendar |          |        |
-| UUID v4 |                            |          |   ✅    |
-| UUID v6 | ✅ <br/> gregorian calendar |    ✅    |        |
-| UUID v7 |     ✅ <br/>unix epoch      |    ✅    |   ✅    |
+| UUID v6 | ✅ <br/> gregorian calendar |    ✅     |        |
+| UUID v7 |     ✅ <br/>unix epoch      |    ✅     |   ✅    |
 
 Implementation based on this [UUID RFC Draft](https://datatracker.ietf.org/doc/html/draft-ietf-uuidrev-rfc4122bis-03)
 
-In addition to UUID, there is also support for [TypeIDs](https://github.com/jetpack-io/typeid). TypeIDs are a modern, 
-type-safe extension of UUIDv7 
+In addition to UUID, there is also support for [TypeIDs](https://github.com/jetpack-io/typeid). TypeIDs are a modern,
+type-safe extension of UUIDv7
 
-##  Quickstart
+_ZIO implementation note:_ Contrary to `uuid4cats-effect`, that we don't provide a UUIDv4 implementation in this lib. ZIO is already providing one
+with `ZIO.randomWith(_.nextUUID)`
 
-To use uuid4cats-effect in an existing SBT project with Scala 2.13 or a later version, add the following dependency to your
-`build.sbt`:
+## Installation
+
+In order to use this library, we need to add the following line in our `build.sbt` file:
 
 ```scala
-libraryDependencies += "ant8e.tech" %% "uuid4cats-effect" % "<version>"
+libraryDependencies += "dev.zio" %% "zio-uuid" % "<version>"
 ```
 
 ## Example
@@ -37,22 +40,47 @@ libraryDependencies += "ant8e.tech" %% "uuid4cats-effect" % "<version>"
 ```scala
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
-import tech.ant8e.uuid4cats.UUIDv6
-import tech.ant8e.uuid4cats.TypeID
+import zio.uuid.UUIDv6
+import zio.uuid.TypeID
 
-val ids = for {
-  generator <- UUIDv6.generator[IO]
-  typeIDGenerator <- TypeID.generator[IO]
-  uuid1 <- generator.uuid
-  uuid2 <- generator.uuid
-  typeid <- typeIDGenerator.typeid("myprefix")
-} yield (uuid1, uuid2, typeid.value)
-
-ids.unsafeRunSync()
-
-val ids: cats.effect.IO[(java.util.UUID, java.util.UUID, String)] = IO(...)
-val res1: (java.util.UUID, java.util.UUID, String) = (1ee22392-7669-6aa0-8000-ca7de6e5d540,1ee22392-766c-61b0-8000-422a97a9dbaa,myprefix_01h5a2ccabe0080m1hrkj0p0qp)
+val ids =
+  (
+    for {
+      uuid1 <- UUIDGenerator.uuid
+      uuid2 <- UUIDGenerator.uuid
+      typeid <- TypeIDGenerator.generate("myprefix")
+    } yield (uuid1, uuid2, typeid.value)
+  ).provideLayers(UUIDGenerator.uuidV7, TypeIDGenerator.live)
 ```
 
-Uniqueness of generated time-based UUIDs is guaranteed when using the same generator. 
-Collisions across generators are theoretically possible although unlikely.    
+Uniqueness of generated time-based UUIDs is guaranteed when using the same generator.
+Collisions across generators are theoretically possible although unlikely.
+
+## Documentation
+
+Learn more on the [zio-uuid homepage](https://zio.dev/zio-uuid)!
+
+## Contributing
+
+For the general guidelines, see ZIO [contributor's guide](https://zio.dev/contributor-guidelines).
+
+## Code of Conduct
+
+See the [Code of Conduct](https://zio.dev/code-of-conduct)
+
+## Support
+
+Come chat with us on [![Badge-Discord]][Link-Discord].
+
+[Badge-Discord]: https://img.shields.io/discord/629491597070827530?logo=discord "chat on discord"
+[Link-Discord]: https://discord.gg/2ccFBr4 "Discord"
+
+## Credits
+
+This library is a fork of [uuid4cats-effect](https://github.com/ant8e/uuid4cats-effect) made by Antoine Comte (https://github.com/ant8e)
+
+## License
+
+[License](LICENSE)
+
+Copyright 2023-2023 Jules Ivanic and the zio-uuid contributors.
