+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.8.2
+sbt.version = 1.9.4
