+++ project/plugins.sbt
@@ -1,5 +1,12 @@
-addSbtPlugin("org.typelevel" % "sbt-typelevel" % "0.4.22")
-addSbtPlugin("org.typelevel" % "sbt-typelevel-site" % "0.4.22")
-addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.13.2")
-addSbtPlugin("org.scalameta" % "sbt-mdoc" % "2.3.7")
+val zioSbtVersion = "0.4.0-alpha.12"
+
+addSbtPlugin("dev.zio" % "zio-sbt-ecosystem" % zioSbtVersion)
+addSbtPlugin("dev.zio" % "zio-sbt-ci"        % zioSbtVersion)
+addSbtPlugin("dev.zio" % "zio-sbt-website"   % zioSbtVersion)
+
+addSbtPlugin("ch.epfl.scala" % "sbt-scalafix"  % "0.11.0")
+addSbtPlugin("org.scalameta" % "sbt-scalafmt"  % "2.5.0")
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.0.8")
+addSbtPlugin("org.typelevel" % "sbt-tpolecat"  % "0.5.0")
+
+resolvers ++= Resolver.sonatypeOssRepos("public")
